+++ lib/screens/Profile/Profile.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
-import '../../theme.dart';
+import '../../widgets/customAppBar.dart';
 import '../Home/HomePage.dart';
 import 'UI components/body.dart';
 import '../../widgets/bottomNavBar.dart';
